[rename]펀딩 내역 조회 DTO 패키지 수정 (#302)
