--- src/main/java/org/kakaoshare/backend/domain/funding/dto/inquiry/FriendFundingInquiryRequest.java
@@ -1,7 +1,0 @@
-package org.kakaoshare.backend.domain.funding.dto.inquiry;
-
-import lombok.Builder;
-import lombok.Getter;
-
-public record FriendFundingInquiryRequest(String friendProviderId) {
-}
