+++ src/main/java/org/kakaoshare/backend/domain/funding/dto/inquiry/request/FriendFundingInquiryRequest.java
@@ -1,0 +1,4 @@
+package org.kakaoshare.backend.domain.funding.dto.inquiry.request;
+
+public record FriendFundingInquiryRequest(String friendProviderId) {
+}
