+++ src/main/java/org/kakaoshare/backend/domain/funding/repository/query/FundingDetailRepositoryCustom.java
@@ -1,7 +1,7 @@
 package org.kakaoshare.backend.domain.funding.repository.query;
 
 import org.kakaoshare.backend.common.vo.date.Date;
-import org.kakaoshare.backend.domain.funding.dto.inquiry.ContributedFundingHistoryResponse;
+import org.kakaoshare.backend.domain.funding.dto.inquiry.response.ContributedFundingHistoryResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
