+++ src/main/java/org/kakaoshare/backend/domain/funding/repository/query/FundingDetailRepositoryCustomImpl.java
@@ -4,15 +4,18 @@ import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.kakaoshare.backend.common.vo.date.Date;
-import org.kakaoshare.backend.domain.funding.dto.inquiry.ContributedFundingHistoryResponse;
 import org.kakaoshare.backend.domain.funding.dto.inquiry.QContributedFundingHistoryDto;
-import org.kakaoshare.backend.domain.funding.dto.inquiry.QContributedFundingHistoryResponse;
+import org.kakaoshare.backend.domain.funding.dto.inquiry.response.ContributedFundingHistoryResponse;
+import org.kakaoshare.backend.domain.funding.dto.inquiry.response.QContributedFundingHistoryResponse;
 import org.kakaoshare.backend.domain.member.entity.QMember;
 import org.kakaoshare.backend.domain.product.dto.QProductDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
-import static org.kakaoshare.backend.common.util.RepositoryUtils.*;
+import static org.kakaoshare.backend.common.util.RepositoryUtils.createOrderSpecifiers;
+import static org.kakaoshare.backend.common.util.RepositoryUtils.eqExpression;
+import static org.kakaoshare.backend.common.util.RepositoryUtils.periodExpression;
+import static org.kakaoshare.backend.common.util.RepositoryUtils.toPage;
 import static org.kakaoshare.backend.domain.funding.entity.QFunding.funding;
 import static org.kakaoshare.backend.domain.funding.entity.QFundingDetail.fundingDetail;
 import static org.kakaoshare.backend.domain.member.entity.QMember.member;
