+++ src/main/java/org/kakaoshare/backend/domain/funding/service/FundingDetailService.java
@@ -4,15 +4,13 @@ import com.querydsl.core.util.StringUtils;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.kakaoshare.backend.common.dto.PageResponse;
-import org.kakaoshare.backend.domain.friend.dto.response.Friend;
 import org.kakaoshare.backend.domain.friend.service.KakaoFriendService;
-import org.kakaoshare.backend.domain.funding.dto.inquiry.FundingContributorResponse;
+import org.kakaoshare.backend.domain.funding.dto.inquiry.response.FundingContributorResponse;
 import org.kakaoshare.backend.domain.funding.dto.inquiry.request.ContributedFundingHistoryRequest;
 import org.kakaoshare.backend.domain.funding.entity.FundingDetail;
 import org.kakaoshare.backend.domain.funding.repository.FundingDetailRepository;
 import org.kakaoshare.backend.domain.funding.vo.FundingHistoryDate;
 import org.kakaoshare.backend.domain.member.dto.oauth.profile.detail.KakaoFriendListDto;
-import org.kakaoshare.backend.domain.member.service.oauth.OAuthWebClientService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
